+++ tdishr/TdiIntrinsic.c
@@ -71,15 +71,6 @@ typedef struct _bounds {
   int u;
 } BOUNDS;
 
-static void free_begin(void* ptr){
-  if (((struct TdiZoneStruct*)ptr)->a_begin) {
-    free(((struct TdiZoneStruct*)ptr)->a_begin);
-    ((struct TdiZoneStruct*)ptr)->a_begin=NULL;
-  }
-}
-#define FREEBEGIN_ON_EXIT() pthread_cleanup_push(free_begin,&TdiRefZone)
-#define FREEBEGIN_NOW()     pthread_cleanup_pop(1)
-
 #define _MOVC3(a,b,c) memcpy(c,b,a)
 extern int TdiFaultHandlerNoFixup();
 extern int Tdi0Decompile();
@@ -196,41 +187,6 @@ static inline void TRACE(int opcode, int narg,
   add(")\n");
 }
 
-static const DESCRIPTOR(newline, "\n");
-static inline void ADD_COMPILE_INFO() {
-  GET_TDITHREADSTATIC_P;
-  static const DESCRIPTOR(compile_err, "%TDI Error compiling region marked by ^\n");
-  struct descriptor_d *message = &TdiThreadStatic_p->TdiIntrinsic_message;
-  if (!TdiRefZone.a_begin || message->length >= MAXMESS)
-    return;
-  // b------x----c----e
-  // '-l_ok-'-xc-'-ce-'
-  char *b = TdiRefZone.a_begin;
-  char *e = TdiRefZone.a_end;
-  char *c = MINMAX(b, TdiRefZone.a_cur, e);
-  char *x = MINMAX(b, b + TdiRefZone.l_ok, c);
-  int xc = MINMAX(0, c-x, MAXLINE);
-  int ce = MINMAX(0, e-c, MAXLINE);
-  int bx = MINMAX(0, x-b, MAXLINE);
-  if (xc+ce > MAXLINE)
-    ce = MINMAX(0, ce, MAXFRAC);
-  if (bx+xc+ce > MAXLINE)
-    bx = MINMAX(0, bx, MAXFRAC);
-  int len = bx+xc+2;
-  struct descriptor marker = { len, DTYPE_T, CLASS_S, memset(malloc(len+1),' ',len) };
-  struct descriptor region = { bx+xc+ce, DTYPE_T, CLASS_S, x-bx };
-  marker.pointer[bx]='^';
-  marker.pointer[0] = marker.pointer[len-1]='\n';
-  if (xc>0)
-    marker.pointer[bx+xc]='^';
-  else if (bx>0)
-    marker.pointer[bx]='^';
-  StrAppend(message,(struct descriptor *)&compile_err);
-  StrAppend(message,(struct descriptor *)&region);
-  StrAppend(message,(struct descriptor *)&marker);
-  free(marker.pointer);
-}
-
 struct _fixed {
   int n;
   char f[256];
@@ -251,7 +207,6 @@ EXPORT int TdiIntrinsic(int opcode, int narg, struct descriptor *list[], struct 
   EMPTYXD(tmp);
   FREEXD_ON_EXIT(&tmp);
   FREEXD_ON_EXIT(out_ptr);
-  FREEBEGIN_ON_EXIT();
   status = MDSplusSUCCESS;
   struct descriptor *dsc_ptr;
   TdiThreadStatic_p->TdiIntrinsic_recursion_count++;
@@ -370,30 +325,16 @@ EXPORT int TdiIntrinsic(int opcode, int narg, struct descriptor *list[], struct 
       goto done;
     status = stat1;
   }
-  /********************************
-  Compiler errors get special help.
-  ********************************/
-  if (opcode == OPC_COMPILE
-  && (status==TdiSYNTAX
-   || status==TdiEXTRANEOUS
-   || status==TdiUNBALANCE
-   || status==TreeNOT_OPEN
-   || status==TreeNNF
-   || status==TdiBOMB))
-    ADD_COMPILE_INFO();
   TRACE(opcode, narg, list, out_ptr);
   if (out_ptr)
     MdsFree1Dx(out_ptr, NULL);
  notmp:MdsFree1Dx(&tmp, NULL);
  done:;
   TdiThreadStatic_p->TdiIntrinsic_recursion_count--;
-  if (!TdiThreadStatic_p->TdiIntrinsic_recursion_count) {
+  if (!TdiThreadStatic_p->TdiIntrinsic_recursion_count)
     TdiThreadStatic_p->TdiIntrinsic_mess_stat = status;
-    free_begin(&TdiRefZone);
-  }
   FREE_CANCEL(&tmp);
   FREE_CANCEL(out_ptr);
-  FREE_CANCEL(a_begin);
   return status;
 }
 EXPORT int _TdiIntrinsic(void** ctx, int opcode, int narg, struct descriptor *list[], struct descriptor_xd *out_ptr){
@@ -432,6 +373,7 @@ int Tdi1Debug(int opcode __attribute__ ((unused)),
     // in order to prepend we need to move the original message into temp desc
     struct descriptor_d oldmsg = *message;
     message->length=0; message->pointer=NULL;
+    static const DESCRIPTOR(newline, "\n");
     StrConcat((struct descriptor *)message, &dmsg, &newline, &oldmsg MDS_END_ARG);
     StrFree1Dx(&oldmsg);
   }
